requirements.txt
Flask
google-generativeai
Pillow
PyMuPDF
python-dotenv
gunicorn